+++ VersionInfo.cs
@@ -3,11 +3,11 @@ using System.Reflection;
 [assembly: AssemblyTitle("FreakShow")]
 [assembly: AssemblyProduct("FreakShow")]
 [assembly: AssemblyCompany("kappiderfreak")]
-[assembly: AssemblyVersion("1.2.9.0")]
-[assembly: AssemblyFileVersion("1.2.9.0")]
-[assembly: AssemblyInformationalVersion("1.2.9")]
+[assembly: AssemblyVersion("1.3.0.0")]
+[assembly: AssemblyFileVersion("1.3.0.0")]
+[assembly: AssemblyInformationalVersion("1.3.0")]
 
 internal static class FreakShowVersion
 {
-    public const string Current = "1.2.9";
+    public const string Current = "1.3.0";
 }